Algunos Arreglos

+++ Segundo-Parcial-EnmanuelPaulino/Segundo-Parcial-EnmanuelPaulino/DAL/Script/Script.sql
@@ -6,7 +6,7 @@ create table Asignaturas
 (
 AsignaturasId int primary key identity,
 Descripcion varchar (max),
-Creditos int
+Creditos decimal
 
 )
 
@@ -14,12 +14,12 @@ Create table Estudiantes (
 EstudianteId int primary key identity,
 FecheIngreso datetime,
 Nombres varchar (max),
-Balance int 
+Balance decimal 
 
 )
 
 create table Inscripcion(
 InscripcionId int primary key identity,
 Fecha datetime,
-Monto int
+Monto decimal
 )
